Created separate package for property sources
