+++ inspectit-ocelot-core/src/main/java/rocks/inspectit/ocelot/core/config/propertysources/file/DirectoryPropertySource.java
@@ -1,4 +1,4 @@
-package rocks.inspectit.ocelot.core.config.filebased;
+package rocks.inspectit.ocelot.core.config.propertysources.file;
 
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
